requirements.txt
celery==4.3.0
channels
daphne
Django
django-celery-beat
django-cors-headers
django-rest-framework
django-timezone-field
djangorestframework
djangorestframework-simplejwt
Pillow
psycopg2
PyJWT
python-dotenv
redis
mysqlclient
asgiref
pytz
sqlparse
gunicorn==20.1.0
dj-database-url
django-environ
uvicorn
Twisted[http2,tls]
pusher